+++ src/Game/world/atmosphere/Environment.js
@@ -1,5 +1,5 @@
 import * as THREE from "three"
-import Game from "../Game"
+import Game from "../../Game"
 
 export default class Environment {
   constructor() {
