+++ src/Game/world/structures/Fences.js
@@ -1,27 +1,26 @@
 import * as THREE from "three"
-import Game from "../Game"
+import Game from "../../Game"
 
-export default class Rocks {
+export default class Fences {
   constructor(config = {}) {
     this.game = new Game()
     this.scene = this.game.scene
     this.resources = this.game.resources
 
     this.modelNames = config.models || [
-      "rocks01Model",
-      "rocks02Model",
-      "rocks03Model",
-      "rocks04Model",
+      "fences01Model",
+      "fences02Model",
+      "fences03Model",
+      "fences04Model",
+      "fences05Model",
     ]
     this.positions = config.positions || []
 
-    this.setRocks()
+    this.setFences()
   }
 
-  setRocks() {
+  setFences() {
     if (this.positions.length === 0) return
-
-    this.meshes = []
 
     this.modelNames.forEach((modelName, typeIndex) => {
       const model = this.resources.items[modelName].scene
@@ -39,7 +38,6 @@ export default class Rocks {
       const instances = this.positions.filter((p) => p.type === typeIndex)
       if (instances.length === 0) return
 
-      // Offset so the bottom of the model sits on the floor (y = 0)
       geometry.computeBoundingBox()
       const yOffset = -geometry.boundingBox.min.y
 
@@ -49,11 +47,11 @@ export default class Rocks {
       const matrix = new THREE.Matrix4()
       for (let i = 0; i < instances.length; i++) {
         const pos = instances[i]
-        const s = pos.scale
+        const s = pos.scale ?? 1
         matrix.compose(
           new THREE.Vector3(pos.x, yOffset * s + (pos.y ?? 0), pos.z),
           new THREE.Quaternion().setFromEuler(
-            new THREE.Euler(0, Math.random() * Math.PI * 2, 0),
+            new THREE.Euler(0, pos.rotation ?? 0, 0),
           ),
           new THREE.Vector3(s, s, s),
         )
@@ -61,7 +59,6 @@ export default class Rocks {
       }
 
       this.scene.add(mesh)
-      this.meshes.push(mesh)
     })
   }
 }
