+++ src/Game/world/structures/Rocks.js
@@ -1,26 +1,27 @@
 import * as THREE from "three"
-import Game from "../Game"
+import Game from "../../Game"
 
-export default class Fences {
+export default class Rocks {
   constructor(config = {}) {
     this.game = new Game()
     this.scene = this.game.scene
     this.resources = this.game.resources
 
     this.modelNames = config.models || [
-      "fences01Model",
-      "fences02Model",
-      "fences03Model",
-      "fences04Model",
-      "fences05Model",
+      "rocks01Model",
+      "rocks02Model",
+      "rocks03Model",
+      "rocks04Model",
     ]
     this.positions = config.positions || []
 
-    this.setFences()
+    this.setRocks()
   }
 
-  setFences() {
+  setRocks() {
     if (this.positions.length === 0) return
+
+    this.meshes = []
 
     this.modelNames.forEach((modelName, typeIndex) => {
       const model = this.resources.items[modelName].scene
@@ -38,6 +39,7 @@ export default class Fences {
       const instances = this.positions.filter((p) => p.type === typeIndex)
       if (instances.length === 0) return
 
+      // Offset so the bottom of the model sits on the floor (y = 0)
       geometry.computeBoundingBox()
       const yOffset = -geometry.boundingBox.min.y
 
@@ -47,11 +49,11 @@ export default class Fences {
       const matrix = new THREE.Matrix4()
       for (let i = 0; i < instances.length; i++) {
         const pos = instances[i]
-        const s = pos.scale ?? 1
+        const s = pos.scale
         matrix.compose(
           new THREE.Vector3(pos.x, yOffset * s + (pos.y ?? 0), pos.z),
           new THREE.Quaternion().setFromEuler(
-            new THREE.Euler(0, pos.rotation ?? 0, 0),
+            new THREE.Euler(0, Math.random() * Math.PI * 2, 0),
           ),
           new THREE.Vector3(s, s, s),
         )
@@ -59,6 +61,7 @@ export default class Fences {
       }
 
       this.scene.add(mesh)
+      this.meshes.push(mesh)
     })
   }
 }
