+++ src/Game/world/vegetation/WillowFoliage.js
@@ -1,7 +1,7 @@
 import * as THREE from "three"
 import { mergeGeometries } from "three/examples/jsm/utils/BufferGeometryUtils.js"
-import vertexShader from "../shaders/willow/vertex.glsl"
-import fragmentShader from "../shaders/willow/fragment.glsl"
+import vertexShader from "../../shaders/willow/vertex.glsl"
+import fragmentShader from "../../shaders/willow/fragment.glsl"
 
 export default class WillowFoliage {
   constructor({
